telemetry-system/services/usb-gps-reader/requirements.txt
pyserial>=3.5
pynmea2>=1.18.0
influxdb-client>=1.36.0
